ml-model/requirements.txt
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.7.2
